Use correct ratio to scale down video canvas while resizing.
Before, we use the height ratio to scale down the video canvas when the
width-height ratio of video is higher than that of body, and vice versa.
This is incorrect because in the previously mentioned case, width needs
to be scaled down more than height (ie. width should be used to
calculate the correct ratio.). The same applies to the other case. The
bug wasn't uncovered for the reason that we had resolution of 1920*1080
which still covers (actually overflows) the screen resolution despite
using incorrect scale ratios.

BUG=chromium:848542
TEST=Tested on CCA 5.1.0 by toggling on-screen keyboard and arbitrarily
resizing the window.

Change-Id: Id13121fe3eeb3193e5036396b577bc0cc8f5600f
Reviewed-on: https://chromium-review.googlesource.com/1086549
Reviewed-by: Sheng-hao Tsao <shenghao@google.com>
Tested-by: Jasmine Chen <lnishan@google.com>
(cherry picked from commit a49c46248035ce116987e870846d4e734f2388e0)
Reviewed-on: https://chromium-review.googlesource.com/1098429
Tested-by: Sheng-hao Tsao <shenghao@google.com>